+++ changelog.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-`4.1.1 (unreleased) <https://github.com/kdeldycke/meta-package-manager/compare/v4.1.0...develop>`_
+`4.2.0 (unreleased) <https://github.com/kdeldycke/meta-package-manager/compare/v4.1.0...develop>`_
 --------------------------------------------------------------------------------------------------
 
 .. note:: This version is not yet released and is under active development.